requirements.txt
asgiref==3.2.3
Django==3.0.3
django-admin-tools==0.9.0
django-model-utils==4.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
Pillow==7.1.1
django-crispy-forms==1.9.0
django-cleanup==4.0.0
